Move shared pytest utility hooks to scalyr_agent.pytest_util module so they can be re-used elsewhere.
